Modify: refetch 실행 전 index 저장을 위한 컴포넌트 분리
